docs(button): update button docs and structure
