+++ .storybook/main.js
@@ -5,8 +5,9 @@ module.exports = {
     '../projects/canopy/src/lib/alert/alert.stories.ts',
     '../projects/canopy/src/lib/brand-icon/brand-icons.stories.ts',
     '../projects/canopy/src/lib/breadcrumb/breadcrumb.stories.ts',
-    '../projects/canopy/src/lib/button/button.stories.ts',
-    '../projects/canopy/src/lib/button/button-group/button-group.stories.ts',
+    '../projects/canopy/src/lib/button/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/button/docs/button.stories.ts',
+    '../projects/canopy/src/lib/button/docs/button-group.stories.ts',
     '../projects/canopy/src/lib/card/card.stories.ts',
     '../projects/canopy/src/lib/carousel/carousel.stories.ts',
     '../projects/canopy/src/lib/data-point/data-point.stories.ts',
