--- projects/canopy/src/lib/button/button-group/button-group.notes.ts
@@ -1,14 +1,0 @@
-export const notes = `
-Provides a container to group together two or more buttons and place them along a single line.
-
-Useful for forms where a 'Submit' and 'Cancel' button are placed side by side, for example.
-
-Import the button module in your module:
-
-~~~js
-@NgModule({
-  ...
-  imports: [ ..., LgButtonModule ],
-})
-~~~
-`;
